migrate crafting funcs from player to character
