+++ src/character.h
@@ -27,13 +27,13 @@
 #include "creature.h"
 #include "cursesdef.h"
 #include "damage.h"
-#include "enums.h"
 #include "enum_int_operators.h"
+#include "enums.h"
 #include "flat_set.h"
 #include "game_constants.h"
 #include "inventory.h"
-#include "item_handling_util.h"
 #include "item.h"
+#include "item_handling_util.h"
 #include "location_ptr.h"
 #include "memory_fast.h"
 #include "pimpl.h"
@@ -62,7 +62,6 @@ class known_magic;
 class player;
 class player_activity;
 class player_morale;
-class recipe_subset;
 class vehicle;
 class monster;
 class weather_manager;
@@ -79,9 +78,23 @@ struct points_left;
 struct trap;
 template <typename E> struct enum_traits;
 
+
+enum class recipe_filter_flags : int;
+class craft_command;
+class recipe;
+class recipe_subset;
+struct requirement_data;
+struct item_comp;
+struct tool_comp;
+
 enum class character_stat : char;
 
 constexpr int MAX_CLAIRVOYANCE = 40;
+
+static const std::string
+DEFAULT_HOTKEYS( "1234567890abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ" );
+
+using recipe_filter = std::function<bool( const recipe &r )>;
 
 enum vision_modes {
     DEBUG_NIGHTVISION,
@@ -232,6 +245,18 @@ struct mountable_status {
     inline bool can_mount() const {
         return mountable && skills && size && carry_weight;
     };
+};
+
+/** @relates ret_val */
+template<>
+struct ret_val<edible_rating>::default_success : public
+    std::integral_constant<edible_rating, edible_rating::edible> {
+};
+
+/** @relates ret_val */
+template<>
+struct ret_val<edible_rating>::default_failure : public
+    std::integral_constant<edible_rating, edible_rating::inedible> {
 };
 
 class Character : public Creature, public location_visitable<Character>
@@ -2145,11 +2170,43 @@ class Character : public Creature, public location_visitable<Character>
         void rem_morale( const morale_type &type );
         void clear_morale();
         bool has_morale_to_read() const;
-        bool has_morale_to_craft() const;
+        bool has_morale_to_craft() const {
+            return true;
+        };
         const inventory &crafting_inventory( bool clear_path );
         const inventory &crafting_inventory( const tripoint &src_pos = tripoint_zero,
                                              int radius = PICKUP_RANGE, bool clear_path = true );
         void invalidate_crafting_inventory();
+
+        double crafting_success_roll( const recipe &making ) const;
+
+        bool can_continue_craft( item &craft );
+
+        void craft_skill_gain( const item &craft, const int &multiplier );
+
+        const requirement_data *select_requirements(
+            const std::vector<const requirement_data *> &, int batch, const inventory &,
+            const std::function<bool( const item & )> &filter ) const;
+        comp_selection<item_comp>
+
+        select_item_component( const std::vector<item_comp> &components,
+                               int batch, inventory &map_inv, bool can_cancel = false,
+                               const std::function<bool( const item & )> &filter = return_true<item>, bool player_inv = true );
+        std::vector<detached_ptr<item>> consume_items( const comp_selection<item_comp> &is, int batch,
+                                     const std::function<bool( const item & )> &filter = return_true<item> );
+        std::vector<detached_ptr<item>> consume_items( map &m, const comp_selection<item_comp> &is,
+                                     int batch,
+                                     const std::function<bool( const item & )> &filter = return_true<item>,
+                                     const tripoint &origin = tripoint_zero, int radius = PICKUP_RANGE );
+        std::vector<detached_ptr<item>> consume_items( const std::vector<item_comp> &components,
+                                     int batch = 1,
+                                     const std::function<bool( const item & )> &filter = return_true<item> );
+        bool craft_consume_tools( item &craft, int mulitplier, bool start_craft );
+        void consume_tools( const comp_selection<tool_comp> &tool, int batch );
+        void consume_tools( map &m, const comp_selection<tool_comp> &tool, int batch,
+                            const tripoint &origin = tripoint_zero, int radius = PICKUP_RANGE );
+        void consume_tools( const std::vector<tool_comp> &tools, int batch = 1,
+                            const std::string &hotkeys = DEFAULT_HOTKEYS );
 
         /** Returns all known recipes. */
         const recipe_subset &get_learned_recipes() const;
@@ -2396,6 +2453,62 @@ class Character : public Creature, public location_visitable<Character>
         void clear_npc_ai_info_cache( npc_ai_info key ) const;
         void set_npc_ai_info_cache( npc_ai_info key, double val ) const;
         std::optional<double> get_npc_ai_info_cache( npc_ai_info key ) const;
+
+    public:
+
+        // Checks crafting inventory for books providing the requested recipe.
+        // Then checks nearby NPCs who could provide it too.
+        // Returns -1 to indicate recipe not found, otherwise difficulty to learn.
+        int has_recipe( const recipe *r, const inventory &crafting_inv,
+                        const std::vector<npc *> &helpers ) const;
+        bool has_recipe_requirements( const recipe &rec ) const;
+
+
+        /** Returns all recipes that are known from the books (either in inventory or nearby). */
+        recipe_subset get_recipes_from_books( const inventory &crafting_inv,
+                                              const recipe_filter &filter = nullptr ) const;
+        /**
+          * Returns all available recipes (from books and npc companions)
+          * @param crafting_inv Current available items to craft
+          * @param helpers List of NPCs that could help with crafting.
+          * @param filter If set, will return only recipes that match the filter (should be much faster).
+          */
+        recipe_subset get_available_recipes( const inventory &crafting_inv,
+                                             const std::vector<npc *> *helpers = nullptr,
+                                             recipe_filter filter = nullptr ) const;
+
+        void make_craft_with_command( const recipe_id &id_to_make, int batch_size, bool is_long = false,
+                                      const tripoint &loc = tripoint_zero );
+        pimpl<craft_command> last_craft;
+
+        recipe_id lastrecipe;
+        int last_batch = 0;
+        itype_id lastconsumed;        //used in crafting.cpp and construction.cpp
+
+        bool studied_all_recipes( const itype &book ) const;
+
+        std::vector<const item *> get_eligible_containers_for_crafting() const;
+        bool check_eligible_containers_for_crafting( const recipe &rec, int batch_size = 1 ) const;
+        bool can_make( const recipe *r, int batch_size = 1 ); // have components?
+        /**
+         * Returns true if the player can start crafting the recipe with the given batch size
+         * The player is not required to have enough tool charges to finish crafting, only to
+         * complete the first step (total / 20 + total % 20 charges)
+         */
+        bool can_start_craft( const recipe *rec, recipe_filter_flags, int batch_size = 1 );
+        bool making_would_work( const recipe_id &id_to_make, int batch_size );
+
+        /** consume components and create an active, in progress craft containing them */
+        item *start_craft( craft_command &command, const tripoint &loc );
+        /**
+         * Start various types of crafts
+         * @param loc the location of the workbench. tripoint_zero indicates crafting from inventory.
+         */
+        void craft( const tripoint &loc = tripoint_zero );
+        void recraft( const tripoint &loc = tripoint_zero );
+        void long_craft( const tripoint &loc = tripoint_zero );
+        void make_craft( const recipe_id &id, int batch_size, const tripoint &loc = tripoint_zero );
+        void make_all_craft( const recipe_id &id, int batch_size, const tripoint &loc = tripoint_zero );
 
 };
