+++ src/player.h
@@ -1,80 +1,25 @@
 #pragma once
 
-#include <climits>
-#include <functional>
-#include <list>
-#include <map>
-#include <memory>
-#include <optional>
 #include <set>
 #include <string>
-#include <type_traits>
-#include <utility>
 #include <vector>
 
-#include "bodypart.h"
-#include "calendar.h"
 #include "character.h"
 #include "character_id.h"
 #include "color.h"
-#include "creature.h"
-#include "cursesdef.h"
 #include "enums.h"
 #include "game_constants.h"
 #include "item.h"
-#include "pimpl.h"
 #include "point.h"
-#include "ret_val.h"
 #include "safe_reference.h"
 #include "string_id.h"
 #include "type_id.h"
 
-class craft_command;
-class effect;
-class faction;
-class inventory;
-class map;
-class npc;
-class recipe;
-struct damage_unit;
-struct requirement_data;
-
-enum class recipe_filter_flags : int;
-struct itype;
-
-static const std::string
-DEFAULT_HOTKEYS( "1234567890abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ" );
-
-class recipe_subset;
-
-enum action_id : int;
 class JsonIn;
 class JsonObject;
 class JsonOut;
-class dispersion_sources;
-struct bionic;
-struct dealt_projectile_attack;
-class profession;
-struct trap;
 
 nc_color encumb_color( int level );
-enum game_message_type : int;
-class vehicle;
-struct item_comp;
-struct tool_comp;
-struct w_point;
-
-using recipe_filter = std::function<bool( const recipe &r )>;
-
-/** @relates ret_val */
-template<>
-struct ret_val<edible_rating>::default_success : public
-    std::integral_constant<edible_rating, edible_rating::edible> {};
-
-/** @relates ret_val */
-template<>
-struct ret_val<edible_rating>::default_failure : public
-    std::integral_constant<edible_rating, edible_rating::inedible> {};
 
 class player : public Character
 {
@@ -126,108 +71,6 @@ class player : public Character
          */
         detached_ptr<item> reduce_charges( item *it, int quantity );
 
-        // Checks crafting inventory for books providing the requested recipe.
-        // Then checks nearby NPCs who could provide it too.
-        // Returns -1 to indicate recipe not found, otherwise difficulty to learn.
-        int has_recipe( const recipe *r, const inventory &crafting_inv,
-                        const std::vector<npc *> &helpers ) const;
-        bool has_recipe_requirements( const recipe &rec ) const;
-
-        bool studied_all_recipes( const itype &book ) const;
-
-        /** Returns all recipes that are known from the books (either in inventory or nearby). */
-        recipe_subset get_recipes_from_books( const inventory &crafting_inv,
-                                              const recipe_filter &filter = nullptr ) const;
-        /**
-          * Returns all available recipes (from books and npc companions)
-          * @param crafting_inv Current available items to craft
-          * @param helpers List of NPCs that could help with crafting.
-          * @param filter If set, will return only recipes that match the filter (should be much faster).
-          */
-        recipe_subset get_available_recipes( const inventory &crafting_inv,
-                                             const std::vector<npc *> *helpers = nullptr,
-                                             recipe_filter filter = nullptr ) const;
-
-        /** For use with in progress crafts */
-        int available_assistant_count( const recipe &rec ) const;
-        /**
-         * Time to craft not including speed multiplier
-         */
-        int base_time_to_craft( const recipe &rec, int batch_size = 1 ) const;
-        /**
-         * Expected time to craft a recipe, with assumption that multipliers stay constant.
-         */
-        int expected_time_to_craft( const recipe &rec, int batch_size = 1, bool in_progress = false ) const;
-        std::vector<const item *> get_eligible_containers_for_crafting() const;
-        bool check_eligible_containers_for_crafting( const recipe &rec, int batch_size = 1 ) const;
-        bool can_make( const recipe *r, int batch_size = 1 );  // have components?
-        /**
-         * Returns true if the player can start crafting the recipe with the given batch size
-         * The player is not required to have enough tool charges to finish crafting, only to
-         * complete the first step (total / 20 + total % 20 charges)
-         */
-        bool can_start_craft( const recipe *rec, recipe_filter_flags, int batch_size = 1 );
-        bool making_would_work( const recipe_id &id_to_make, int batch_size );
-
-        /**
-         * Start various types of crafts
-         * @param loc the location of the workbench. tripoint_zero indicates crafting from inventory.
-         */
-        void craft( const tripoint &loc = tripoint_zero );
-        void recraft( const tripoint &loc = tripoint_zero );
-        void long_craft( const tripoint &loc = tripoint_zero );
-        void make_craft( const recipe_id &id, int batch_size, const tripoint &loc = tripoint_zero );
-        void make_all_craft( const recipe_id &id, int batch_size, const tripoint &loc = tripoint_zero );
-        /** consume components and create an active, in progress craft containing them */
-        item *start_craft( craft_command &command, const tripoint &loc );
-        /**
-         * Calculate a value representing the success of the player at crafting the given recipe,
-         * taking player skill, recipe difficulty, npc helpers, and player mutations into account.
-         * @param making the recipe for which to calculate
-         * @return a value >= 0.0 with >= 1.0 representing unequivocal success
-         */
-        double crafting_success_roll( const recipe &making ) const;
-        /**
-         * Check if the player meets the requirements to continue the in progress craft and if
-         * unable to continue print messages explaining the reason.
-         * If the craft is missing components due to messing up, prompt to consume new ones to
-         * allow the craft to be continued.
-         * @param craft the currently in progress craft
-         * @return if the craft can be continued
-         */
-        bool can_continue_craft( item &craft );
-        /**
-         * Handle skill gain for player and followers during crafting
-         * @param craft the currently in progress craft
-         * @param multiplier what factor to multiply the base skill gain by.  This is used to apply
-         * multiple steps of incremental skill gain simultaneously if needed.
-         */
-        void craft_skill_gain( const item &craft, const int &multiplier );
-
-        const requirement_data *select_requirements(
-            const std::vector<const requirement_data *> &, int batch, const inventory &,
-            const std::function<bool( const item & )> &filter ) const;
-        comp_selection<item_comp>
-        select_item_component( const std::vector<item_comp> &components,
-                               int batch, inventory &map_inv, bool can_cancel = false,
-                               const std::function<bool( const item & )> &filter = return_true<item>, bool player_inv = true );
-        std::vector<detached_ptr<item>> consume_items( const comp_selection<item_comp> &is, int batch,
-                                     const std::function<bool( const item & )> &filter = return_true<item> );
-        std::vector<detached_ptr<item>> consume_items( map &m, const comp_selection<item_comp> &is,
-                                     int batch,
-                                     const std::function<bool( const item & )> &filter = return_true<item>,
-                                     const tripoint &origin = tripoint_zero, int radius = PICKUP_RANGE );
-        std::vector<detached_ptr<item>> consume_items( const std::vector<item_comp> &components,
-                                     int batch = 1,
-                                     const std::function<bool( const item & )> &filter = return_true<item> );
-        /** Consume tools for the next multiplier * 5% progress of the craft */
-        bool craft_consume_tools( item &craft, int mulitplier, bool start_craft );
-        void consume_tools( const comp_selection<tool_comp> &tool, int batch );
-        void consume_tools( map &m, const comp_selection<tool_comp> &tool, int batch,
-                            const tripoint &origin = tripoint_zero, int radius = PICKUP_RANGE );
-        void consume_tools( const std::vector<tool_comp> &tools, int batch = 1,
-                            const std::string &hotkeys = DEFAULT_HOTKEYS );
-
         // ---------------VALUES-----------------
         tripoint view_offset;
         // Relative direction of a grab, add to posx, posy to get the coordinates of the grabbed thing.
@@ -245,14 +88,6 @@ class player : public Character
         bool manual_examine = false;
         vproto_id starting_vehicle = vproto_id::NULL_ID();
         std::vector<mtype_id> starting_pets;
-
-        void make_craft_with_command( const recipe_id &id_to_make, int batch_size, bool is_long = false,
-                                      const tripoint &loc = tripoint_zero );
-        pimpl<craft_command> last_craft;
-
-        recipe_id lastrecipe;
-        int last_batch = 0;
-        itype_id lastconsumed;        //used in crafting.cpp and construction.cpp
 
         std::set<character_id> follower_ids;
