+++ src/requirements.cpp
@@ -28,7 +28,7 @@
 #include "locations.h"
 #include "make_static.h"
 #include "output.h"
-#include "player.h"
+#include "Character.h"
 #include "point.h"
 #include "salvage.h"
 #include "string_formatter.h"
@@ -1558,14 +1558,14 @@ std::vector<const requirement_data *> deduped_requirement_data::feasible_alterna
 }
 
 const requirement_data *deduped_requirement_data::select_alternative(
-    player &crafter, const std::function<bool( const item & )> &filter, int batch,
+    Character &crafter, const std::function<bool( const item & )> &filter, int batch,
     cost_adjustment flags ) const
 {
     return select_alternative( crafter, crafter.crafting_inventory(), filter, batch, flags );
 }
 
 const requirement_data *deduped_requirement_data::select_alternative(
-    player &crafter, const inventory &inv, const std::function<bool( const item & )> &filter,
+    Character &crafter, const inventory &inv, const std::function<bool( const item & )> &filter,
     int batch, cost_adjustment flags ) const
 {
     const std::vector<const requirement_data *> all_reqs =
